rename language_model_transcribe -> transcribe
it contains more than one transcribe method now
