+++ gentle/transcribe.py
