+++ serve.py
@@ -12,7 +12,7 @@ import sys
 import uuid
 
 from gentle.paths import get_binary, get_resource, get_datadir
-from gentle.language_model_transcribe import lm_transcribe, write_csv
+from gentle.transcribe import lm_transcribe, write_csv
 
 DATADIR = get_datadir('webdata')
