+++ tests/data/update_golden_master.sh
@@ -2,4 +2,4 @@
 
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
-python $DIR/../../gentle/language_model_transcribe.py $DIR/lucier.mp3 $DIR/lucier.txt $DIR/lucier_golden.json
+python $DIR/../../gentle/transcribe.py $DIR/lucier.mp3 $DIR/lucier.txt $DIR/lucier_golden.json
